perf: pack availability enrichment fields into existing padding
FortLookup is held per resident fort, so the two int16 enrichment
fields cost a measured 8 bytes each after alignment (184 -> 192).
Both values are tiny enums: as int8 placed in RaidLevel's pad byte and
the pad before ShowcaseExpiry, unsafe.Sizeof returns 184 again — the
enrichment now adds zero bytes per fort.

+++ decoder/api_pokestop_available.go
@@ -47,7 +47,7 @@ type ApiPokestopShowcaseAvailable struct {
 	PokemonId       *int16 `json:"pokemon_id" doc:"Showcase focus pokemon id; null for a type-based showcase"`
 	Form            *int16 `json:"form" doc:"Showcase focus pokemon form (0 is a valid form); null for a type-based showcase"`
 	TypeId          *int8  `json:"type_id" doc:"Showcase focus pokemon type id (type-based showcases); null for a pokemon-based showcase"`
-	RankingStandard int16  `json:"ranking_standard" doc:"Ranking standard of the showcase contest; 0 when unknown"`
+	RankingStandard int8   `json:"ranking_standard" doc:"Ranking standard of the showcase contest; 0 when unknown"`
 }
 
 // ApiAvailablePokestops is the whole-instance snapshot served by
